[WebAudio] Block AudioContext until the HRTF database is fully loaded
The Raspberry Pi has only one CPU core and keeping the audio context
active while the HRTF db loads is seriously slowing down the HRTF db
thread. In that configuration it takes about 23 seconds to load.

Blocking the AudioContext while the HRTF db thread is running allows a
faster loading of the db (in about 13 seconds). The biggest drawback
though is that the entire WebProcess is blocked during that time.